Remove ActorDispatcher table, refactor DispatcherType usage, and clean up imports for readability.
