+++ src/meta/model/migration/src/m20250509_102041_remove_dispatcher.rs
@@ -1,0 +1,21 @@
+use sea_orm_migration::prelude::*;
+
+#[derive(DeriveMigrationName)]
+pub struct Migration;
+
+#[async_trait::async_trait]
+impl MigrationTrait for Migration {
+    async fn up(&self, manager: &SchemaManager) -> Result<(), DbErr> {
+        crate::drop_tables!(manager, ActorDispatcher);
+        Ok(())
+    }
+
+    async fn down(&self, _manager: &SchemaManager) -> Result<(), DbErr> {
+        Ok(())
+    }
+}
+
+#[derive(DeriveIden)]
+enum ActorDispatcher {
+    Table,
+}
